batch/requirements.txt
mysql-connector-python==8.0.20
boto3==1.17.44
botocore==1.20.44
python-dotenv==0.15.0
